chore(developer): kmcmplib - rename kmcmpdll.h to reduce confusion
