+++ developer/src/kmcmplib/src/CheckNCapsConsistency.cpp
@@ -3,7 +3,7 @@
 
 #include <compfile.h>
 #include <comperr.h>
-#include <kmcmpdll.h>
+#include <kmcmplib.h>
 #include "CharToKeyConversion.h"
 
 /**
