+++ developer/src/kmcmplib/src/versioning.cpp
@@ -1,7 +1,7 @@
 #include "pch.h"
 #include "compfile.h"
 #include <comperr.h>
-#include "kmcmpdll.h"
+#include "kmcmplib.h"
 #include "versioning.h"
 
 //namespace kmcmp {
